Add files via upload
This code first logs into a website (the "headers" variable is a dictionary that contains the HTTP headers that are sent to the server with an HTTP request. In this case, the headers dictionary has a single key-value pair, with the key being "User-Agent" and the value being "MacBook".. It does this by navigating to the login page and attempting to find the token that is required to log in. It then creates a dictionary with this token and the login form and submits the data to the login page to log in to the system. After successfully logging in, the code navigates to the page with quotes and performs the necessary data extraction from the page. This code has a pause between each request to avoid overloading the server.  It then creates a generator function base that iterates over the text and author lists and yields tuples of quote text and author. Finally, the writer function is called and passed base as an argument. The writer function then creates an Excel workbook and writes the quotes and authors to the work.